Extract MigrationResource logic into smaller units
